+++ cmd/boulder/main.go
@@ -12,9 +12,9 @@ import (
 	_ "github.com/letsencrypt/boulder/cmd/boulder-observer"
 	_ "github.com/letsencrypt/boulder/cmd/boulder-publisher"
 	_ "github.com/letsencrypt/boulder/cmd/boulder-ra"
-	_ "github.com/letsencrypt/boulder/cmd/boulder-remoteva"
 	_ "github.com/letsencrypt/boulder/cmd/boulder-sa"
-	_ "github.com/letsencrypt/boulder/cmd/boulder-va"
+	_ "github.com/letsencrypt/boulder/cmd/boulder-va/rva"
+	_ "github.com/letsencrypt/boulder/cmd/boulder-va/va"
 	_ "github.com/letsencrypt/boulder/cmd/boulder-wfe2"
 	_ "github.com/letsencrypt/boulder/cmd/caa-log-checker"
 	_ "github.com/letsencrypt/boulder/cmd/cert-checker"
